+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added [`DateOnlyRange`](/src/Candoumbe.MiscUtilities/Types/DateOnlyRange.cs) type
+- Added [`TimeOnlyRange`](/src/Candoumbe.MiscUtilities/Types/TimeOnlyRange.cs) type
+- Added [`DateTimeRange`](/src/Candoumbe.MiscUtilities/Types/DateTimeRange.cs) type
+- Added generic [`Range<T>`](/src/Candoumbe.MiscUtilities/Types/DateTimeRange.cs) type
 
 ## [0.8.0] / 2022-01-26
 - Added `SortBy` extension method for `IEnumerable<T>` type